refactor: JPA 엔티티 매핑 설정
