--- src/main/java/roomescape/domain/Reservation.java
@@ -1,77 +1,0 @@
-package roomescape.domain;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import roomescape.exception.InvalidReservationException;
-
-public class Reservation {
-
-    private final Long id;
-    private final Member member;
-    private final LocalDate date;
-    private final ReservationTime time;
-    private final Theme theme;
-
-    public Reservation(Long id, Member member, LocalDate date, ReservationTime time, Theme theme) {
-        validate(member, date, time, theme);
-        this.id = id;
-        this.member = member;
-        this.date = date;
-        this.time = time;
-        this.theme = theme;
-    }
-
-    public Reservation(Member member, LocalDate date, ReservationTime time, Theme theme) {
-        this(null, member, date, time, theme);
-    }
-
-    private void validate(Member member, LocalDate date, ReservationTime time, Theme theme) {
-        if (member == null) {
-            throw new InvalidReservationException("예약자는 비어있을 수 없습니다");
-        }
-        if (date == null || time == null) {
-            throw new InvalidReservationException("시간은 비어있을 수 없습니다.");
-        }
-        if (theme == null) {
-            throw new InvalidReservationException("테마는 비어있을 수 없습니다.");
-        }
-    }
-
-    public boolean isBefore(LocalDateTime compareDateTime) {
-        LocalDateTime reservationDateTime = LocalDateTime.of(date, getStartAt());
-        return reservationDateTime.isBefore(compareDateTime);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public String getName() {
-        return member.getName();
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public ReservationTime getReservationTime() {
-        return time;
-    }
-
-    public LocalTime getStartAt() {
-        return time.getTime();
-    }
-
-    public Theme getTheme() {
-        return theme;
-    }
-
-    public String getThemeName() {
-        return theme.getName();
-    }
-}
