--- src/main/java/roomescape/domain/Theme.java
@@ -1,68 +1,0 @@
-package roomescape.domain;
-
-import java.util.Objects;
-import roomescape.exception.InvalidThemeException;
-
-public class Theme {
-
-    private final Long id;
-    private final String name;
-    private final String description;
-    private final String thumbnail;
-
-    public Theme(Long id, String name, String description, String thumbnail) {
-        validate(name, description, thumbnail);
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.thumbnail = thumbnail;
-    }
-
-    public Theme(String name, String description, String thumbnail) {
-        this(null, name, description, thumbnail);
-    }
-
-    private void validate(String name, String description, String thumbnail) {
-        if (name == null || name.isBlank()) {
-            throw new InvalidThemeException("테마 이름은 비어있을 수 없습니다.");
-        }
-        if (description == null) {
-            throw new InvalidThemeException("설명은 비어있을 수 없습니다.");
-        }
-        if (thumbnail == null) {
-            throw new InvalidThemeException("썸네일은 비어있을 수 없습니다.");
-        }
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getThumbnail() {
-        return thumbnail;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Theme theme = (Theme) o;
-        return Objects.equals(id, theme.id) && Objects.equals(name, theme.name)
-                && Objects.equals(description, theme.description) && Objects.equals(thumbnail,
-                theme.thumbnail);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, description, thumbnail);
-    }
-}
