+++ src/main/java/roomescape/entity/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.entity;
 
 import roomescape.exception.InvalidMemberException;
