+++ src/main/java/roomescape/entity/ThemeRanking.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.entity;
 
 import java.util.Comparator;
 import java.util.HashMap;
@@ -7,6 +7,7 @@ import java.util.Map;
 import roomescape.entity.Reservation;
 import roomescape.entity.Theme;
 
+// TODO: repo JPA 적용 후, 제거
 public class ThemeRanking {
 
     private static final int RANKING_SIZE = 10;
