+++ src/main/java/roomescape/repository/jdbc/JdbcThemeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 import roomescape.repository.ThemeRepository;
 
 @Repository
